pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crossword-builder"
version = "0.1.0"
description = "Educational crossword puzzle generator from source material"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.9"
authors = [
    {name = "Patrick McKnight", email = "pem725@gmail.com"}
]
keywords = ["crossword", "puzzle", "education", "vocabulary"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

[project.optional-dependencies]
render = [
    "Pillow>=9.0",
    "fpdf2>=2.7",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "Pillow>=9.0",
    "fpdf2>=2.7",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
